+++ CHANGELOG.md
@@ -1,3 +1,8 @@
 # Network Location Changelog
 
+## [1.0.0] - 2023-05-18
+- Added ability to run location switch with AppleScript and admin privileges
+- Updated command icon
+- Implemented core functionality to switch between network locations
+
 ## [Initial Version] - 2023-05-17